pyproject.toml
[tool.poetry]
name = "wellets_cli"
version = "1.4.0"
description = "wellets-cli is the command line interface for Wellets, a crypto-oriented financial management tool that allows you to keep under control your money."
authors = ["lparolari"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
pydantic = "^1.9.1"
click = "^8.1.3"
tabulate = "^0.8.10"
python-dateutil = "^2.8.2"
inquirerpy = "^0.3.4"
numpy = "^1.23.5"
matplotlib = "^3.6.2"
dynaconf = "^3.1.12"

[tool.poetry.dev-dependencies]
coverage = "^6.4.2"
pytest = "^7.1.2"
flake8 = "^5.0.4"
black = "^22.6.0"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
codecov = "^2.1.12"
mypy = "^0.971"
gitchangelog = "^3.0.4"
mkdocs = "^1.3.1"
types-python-dateutil = "^2.8.19"
types-tabulate = "^0.8.11"
types-requests = "^2.28.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wellets_cli = 'wellets_cli.__main__:main'
